将index、len、size等变量的定义，都配置成size_t而不是uint32_t，至于列表对应的负数索引，先按照int吧，后续有必要再修改成ssize_t

+++ demo/demo_deque.c
@@ -12,10 +12,10 @@
 
 static void demo_deque_num(void)
 {
-    uint32_t i = 0;
+    size_t i = 0;
     int data[] = { 1,2,3,4,5,6,7,8,9,10 };
     int temp = 0;
-    uint32_t len = sizeof(data) / sizeof(data[0]);
+    size_t len = sizeof(data) / sizeof(data[0]);
 
     deque_t deque = deque_new(sizeof(int));
     deque->print_obj = print_num;
@@ -181,14 +181,14 @@ static void demo_deque_num(void)
 
 static void demo_deque_struct(void)
 {
-    uint32_t i = 0;
+    size_t i = 0;
     struct _student data[] = {
         "zhao", 1001, "qian", 1002, "sun",   1003, "li",   1004,
         "zhou", 1005, "wu",   1006, "zheng", 1007, "wang", 1008,
         "feng", 1009, "cheng",1010,
     };
     struct _student temp = {0};
-    uint32_t len = sizeof(data) / sizeof(data[0]);
+    size_t len = sizeof(data) / sizeof(data[0]);
 
     deque_t deque = deque_new(sizeof(struct _student));
     deque->print_obj = print_struct;
