+++ demo/demo_queue.c
@@ -12,10 +12,10 @@
 
 static void demo_queue_num(void)
 {
-    uint32_t i = 0;
+    size_t i = 0;
     int data[] = { 1,2,3,4,5,6,7,8,9,10 };
     int temp = 0;
-    uint32_t len = sizeof(data) / sizeof(data[0]);
+    size_t len = sizeof(data) / sizeof(data[0]);
 
     queue_t queue = queue_new(sizeof(int));
     queue->print_obj = print_num;
@@ -96,10 +96,10 @@ static void demo_queue_num(void)
 
 static void demo_queue_char(void)
 {
-    uint32_t i = 0;
+    size_t i = 0;
     char data[] = "abcdefghijk";
     char temp = 0;
-    uint32_t len = sizeof(data) / sizeof(data[0]) - 1;
+    size_t len = sizeof(data) / sizeof(data[0]) - 1;
 
     queue_t queue = queue_new2(sizeof(char), 10);
     queue->print_obj = print_char;
@@ -205,13 +205,13 @@ static void demo_queue_char(void)
 
 static void demo_queue_struct(void)
 {
-    uint32_t i = 0;
+    size_t i = 0;
     struct _student data[] = {
         {"zhao", 1001},{"qian", 1002}, {"sun", 1003}, {"li", 1004},
         "zhou", 1005, "wu", 1006, "zheng", 1007, "wang", 1008,
     };
     struct _student temp;
-    uint32_t len = sizeof(data) / sizeof(data[0]) - 1;
+    size_t len = sizeof(data) / sizeof(data[0]) - 1;
 
     queue_t queue = queue_new(sizeof(struct _student));
     queue->print_obj = print_struct;
